Migrate to MCP Framework, refactor tools, and update project structure
