+++ mcp-elizaos-data/.gitignore
@@ -1,19 +1,7 @@
-node_modules/
-build/
-dist/
-*.tgz
+node_modules
+dist
 .env
-.env.local
-.env.development.local
-.env.test.local
-.env.production.local
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
+logs
 .DS_Store
-Thumbs.db
-.vscode/
-.idea/
-*.swp
-*.swo
-*~
+.idea
+.vscode
