--- mcp-elizaos-data/.npmignore
@@ -1,16 +1,0 @@
-src/
-*.ts
-!build/**/*.d.ts
-tsconfig.json
-.git
-.gitignore
-node_modules/
-test/
-*.test.js
-*.spec.js
-.env
-.env.*
-coverage/
-.nyc_output/
-.DS_Store
-Thumbs.db
