+++ mcp-elizaos-data/package.json
@@ -2,17 +2,20 @@
   "name": "mcp-elizaos-data",
   "version": "1.0.0",
   "description": "MCP server that exposes ElizaOS data endpoints for real-time ecosystem information",
-  "main": "build/index.js",
   "type": "module",
   "bin": {
-    "mcp-elizaos-data": "build/index.js"
+    "mcp-elizaos-data": "./dist/index.js"
   },
+  "files": [
+    "dist",
+    "README.md"
+  ],
   "scripts": {
-    "build": "tsc",
-    "start": "node build/index.js",
-    "dev": "tsx src/index.ts",
-    "prepare": "npm run build",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "build": "tsc && mcp-build",
+    "watch": "tsc --watch",
+    "start": "node dist/index.js",
+    "validate": "mcp validate",
+    "prepare": "npm run build"
   },
   "keywords": [
     "mcp",
@@ -32,22 +35,15 @@
     "directory": "mcp-elizaos-data"
   },
   "homepage": "https://elizaos.github.io",
-  "files": [
-    "build/**/*",
-    "README.md",
-    "package.json"
-  ],
-  "engines": {
-    "node": ">=18.0.0"
-  },
   "dependencies": {
-    "@modelcontextprotocol/sdk": "^1.0.0",
-    "axios": "^1.6.0",
-    "zod": "^3.22.0"
+    "axios": "^1.11.0",
+    "mcp-framework": "^0.2.2"
   },
   "devDependencies": {
-    "@types/node": "^20.0.0",
-    "tsx": "^4.0.0",
-    "typescript": "^5.0.0"
+    "@types/node": "^20.11.24",
+    "typescript": "^5.3.3"
+  },
+  "engines": {
+    "node": ">=18.19.0"
   }
 }
