--- mcp-elizaos-data/src/tools/get-plugin-stats.ts
@@ -1,165 +1,0 @@
-import { z } from "zod";
-import { fetchData } from "../http-client.js";
-
-// Input schema
-export const GetPluginStatsArgsSchema = z.object({
-  plugin_name: z.string().optional().describe("Specific plugin name to get stats for"),
-  include_details: z.boolean().optional().default(false).describe("Include detailed statistics"),
-});
-
-// Output schema
-const PluginStatSchema = z.object({
-  name: z.string().describe("Plugin name"),
-  downloads: z.number().optional().describe("Number of downloads"),
-  stars: z.number().optional().describe("Number of stars"),
-  forks: z.number().optional().describe("Number of forks"),
-  contributors: z.number().optional().describe("Number of contributors"),
-  last_updated: z.string().optional().describe("Last update timestamp"),
-  version: z.string().optional().describe("Latest version"),
-});
-
-const PluginStatsResponseSchema = z.object({
-  total_plugins: z.number().describe("Total number of plugins"),
-  plugins: z.array(PluginStatSchema).describe("Array of plugin statistics"),
-  updated_at: z.string().describe("When the data was last updated"),
-});
-
-export type GetPluginStatsArgs = z.infer<typeof GetPluginStatsArgsSchema>;
-export type PluginStatsResponse = z.infer<typeof PluginStatsResponseSchema>;
-
-// Tool schema for MCP
-export const getPluginStatsSchema = {
-  name: "get_plugin_stats",
-  description: "Fetch plugin statistics from ElizaOS data including downloads, stars, and other metrics",
-  inputSchema: GetPluginStatsArgsSchema,
-};
-
-export async function getPluginStatsHandler(args: any) {
-  const validatedArgs = GetPluginStatsArgsSchema.parse(args);
-  
-  try {
-    // Try multiple potential endpoints for plugin stats
-    const endpoints = [
-      "/elizaos-plugins_registry/stats.json",
-      "/summaries/plugin-stats.json",
-      "/dump/plugin-stats.json",
-    ];
-
-    let result = null;
-    let endpoint_used = "";
-
-    for (const endpoint of endpoints) {
-      const response = await fetchData<any>(endpoint);
-      if (response.success && response.data) {
-        result = response.data;
-        endpoint_used = endpoint;
-        break;
-      }
-    }
-
-    if (!result) {
-      // Create mock data structure if no real endpoint is available
-      const mockData = {
-        total_plugins: 15,
-        plugins: [
-          {
-            name: "plugin-evm",
-            downloads: 1250,
-            stars: 45,
-            forks: 12,
-            contributors: 8,
-            last_updated: "2024-01-15T10:30:00Z",
-            version: "1.2.0"
-          },
-          {
-            name: "plugin-solana", 
-            downloads: 980,
-            stars: 38,
-            forks: 9,
-            contributors: 6,
-            last_updated: "2024-01-14T15:45:00Z",
-            version: "1.1.5"
-          },
-          {
-            name: "plugin-twitter",
-            downloads: 2100,
-            stars: 67,
-            forks: 23,
-            contributors: 15,
-            last_updated: "2024-01-16T09:20:00Z",
-            version: "2.0.1"
-          }
-        ],
-        updated_at: new Date().toISOString(),
-        source: "Generated from available plugin directories"
-      };
-
-      // Filter by plugin name if specified
-      if (validatedArgs.plugin_name) {
-        const filteredPlugins = mockData.plugins.filter(p => 
-          p.name.toLowerCase().includes(validatedArgs.plugin_name!.toLowerCase())
-        );
-        mockData.plugins = filteredPlugins;
-        mockData.total_plugins = filteredPlugins.length;
-      }
-
-      return {
-        content: [
-          {
-            type: "text",
-            text: JSON.stringify(mockData, null, 2),
-          },
-        ],
-      };
-    }
-
-    // Process the actual data if available
-    let processedData = result;
-    
-    if (validatedArgs.plugin_name) {
-      // Filter by plugin name if specified
-      if (processedData.plugins && Array.isArray(processedData.plugins)) {
-        processedData.plugins = processedData.plugins.filter((p: any) =>
-          p.name?.toLowerCase().includes(validatedArgs.plugin_name!.toLowerCase())
-        );
-        processedData.total_plugins = processedData.plugins.length;
-      }
-    }
-
-    if (!validatedArgs.include_details && processedData.plugins) {
-      // Remove detailed fields if not requested
-      processedData.plugins = processedData.plugins.map((p: any) => ({
-        name: p.name,
-        downloads: p.downloads,
-        stars: p.stars,
-        version: p.version,
-      }));
-    }
-
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify({
-            ...processedData,
-            source_endpoint: endpoint_used,
-          }, null, 2),
-        },
-      ],
-    };
-  } catch (error: any) {
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify({
-            error: "Failed to fetch plugin stats",
-            message: error.message,
-            timestamp: new Date().toISOString(),
-          }, null, 2),
-        },
-      ],
-      isError: true,
-    };
-  }
-}
