--- mcp-elizaos-data/src/tools/get-repository-data.ts
@@ -1,237 +1,0 @@
-import { z } from "zod";
-import { fetchData } from "../http-client.js";
-
-// Input schema
-export const GetRepositoryDataArgsSchema = z.object({
-  repository_name: z.string().optional().describe("Specific repository name to get data for"),
-  include_stats: z.boolean().optional().default(true).describe("Include repository statistics"),
-  include_contributors: z.boolean().optional().default(false).describe("Include contributor information"),
-});
-
-// Output schema
-const ContributorSchema = z.object({
-  login: z.string().describe("GitHub username"),
-  contributions: z.number().describe("Number of contributions"),
-  avatar_url: z.string().optional().describe("Avatar URL"),
-});
-
-const RepositoryDataSchema = z.object({
-  name: z.string().describe("Repository name"),
-  full_name: z.string().describe("Full repository name (owner/repo)"),
-  description: z.string().optional().describe("Repository description"),
-  stars: z.number().optional().describe("Number of stars"),
-  forks: z.number().optional().describe("Number of forks"),
-  open_issues: z.number().optional().describe("Number of open issues"),
-  language: z.string().optional().describe("Primary programming language"),
-  created_at: z.string().optional().describe("Repository creation date"),
-  updated_at: z.string().optional().describe("Last update date"),
-  contributors: z.array(ContributorSchema).optional().describe("Repository contributors"),
-  topics: z.array(z.string()).optional().describe("Repository topics/tags"),
-});
-
-const RepositoryDataResponseSchema = z.object({
-  total_repositories: z.number().describe("Total number of repositories"),
-  repositories: z.array(RepositoryDataSchema).describe("Array of repository data"),
-  updated_at: z.string().describe("When the data was last updated"),
-});
-
-export type GetRepositoryDataArgs = z.infer<typeof GetRepositoryDataArgsSchema>;
-export type RepositoryDataResponse = z.infer<typeof RepositoryDataResponseSchema>;
-
-// Tool schema for MCP
-export const getRepositoryDataSchema = {
-  name: "get_repository_data",
-  description: "Fetch repository metadata and statistics from ElizaOS data including stars, forks, and contributors",
-  inputSchema: GetRepositoryDataArgsSchema,
-};
-
-export async function getRepositoryDataHandler(args: any) {
-  const validatedArgs = GetRepositoryDataArgsSchema.parse(args);
-  
-  try {
-    // Try multiple potential endpoints for repository data
-    const endpoints = [
-      "/elizaos_elizaos.github.io/repository-data.json",
-      "/elizaos_eliza/metadata.json",
-      "/summaries/repository-summary.json",
-      "/dump/repositories.json",
-    ];
-
-    let result = null;
-    let endpoint_used = "";
-
-    for (const endpoint of endpoints) {
-      const response = await fetchData<any>(endpoint);
-      if (response.success && response.data) {
-        result = response.data;
-        endpoint_used = endpoint;
-        break;
-      }
-    }
-
-    if (!result) {
-      // Create mock data structure if no real endpoint is available
-      const mockData = {
-        total_repositories: 4,
-        repositories: [
-          {
-            name: "eliza",
-            full_name: "elizaos/eliza",
-            description: "AI agent framework for autonomous AI agents",
-            stars: 2847,
-            forks: 612,
-            open_issues: 89,
-            language: "TypeScript",
-            created_at: "2024-01-01T00:00:00Z",
-            updated_at: "2024-01-16T12:00:00Z",
-            contributors: validatedArgs.include_contributors ? [
-              { login: "shawmakesmagic", contributions: 245, avatar_url: "https://github.com/shawmakesmagic.png" },
-              { login: "elizaos-contributor", contributions: 156, avatar_url: "https://github.com/elizaos-contributor.png" },
-              { login: "ai-developer", contributions: 89, avatar_url: "https://github.com/ai-developer.png" }
-            ] : undefined,
-            topics: ["ai", "agent", "autonomous", "typescript", "eliza"]
-          },
-          {
-            name: "elizaos.github.io",
-            full_name: "elizaos/elizaos.github.io",
-            description: "Official ElizaOS website and documentation",
-            stars: 124,
-            forks: 34,
-            open_issues: 12,
-            language: "JavaScript",
-            created_at: "2024-01-15T00:00:00Z",
-            updated_at: "2024-01-16T10:30:00Z",
-            contributors: validatedArgs.include_contributors ? [
-              { login: "docs-maintainer", contributions: 67, avatar_url: "https://github.com/docs-maintainer.png" },
-              { login: "web-developer", contributions: 43, avatar_url: "https://github.com/web-developer.png" }
-            ] : undefined,
-            topics: ["documentation", "website", "nextjs", "elizaos"]
-          },
-          {
-            name: "plugins",
-            full_name: "elizaos/plugins",
-            description: "Official plugin repository for ElizaOS",
-            stars: 456,
-            forks: 123,
-            open_issues: 23,
-            language: "TypeScript",
-            created_at: "2024-01-10T00:00:00Z",
-            updated_at: "2024-01-16T14:20:00Z",
-            contributors: validatedArgs.include_contributors ? [
-              { login: "plugin-dev", contributions: 134, avatar_url: "https://github.com/plugin-dev.png" },
-              { login: "extension-author", contributions: 98, avatar_url: "https://github.com/extension-author.png" }
-            ] : undefined,
-            topics: ["plugins", "extensions", "elizaos", "typescript"]
-          },
-          {
-            name: "auto.fun",
-            full_name: "elizaos/auto.fun",
-            description: "Autonomous agent platform powered by ElizaOS",
-            stars: 789,
-            forks: 201,
-            open_issues: 34,
-            language: "TypeScript",
-            created_at: "2024-01-05T00:00:00Z",
-            updated_at: "2024-01-16T16:45:00Z",
-            contributors: validatedArgs.include_contributors ? [
-              { login: "platform-dev", contributions: 178, avatar_url: "https://github.com/platform-dev.png" },
-              { login: "auto-engineer", contributions: 145, avatar_url: "https://github.com/auto-engineer.png" }
-            ] : undefined,
-            topics: ["autonomous", "platform", "ai", "elizaos"]
-          }
-        ],
-        updated_at: new Date().toISOString(),
-        source: "Generated from available repository directories"
-      };
-
-      // Filter by repository name if specified
-      if (validatedArgs.repository_name) {
-        const filteredRepos = mockData.repositories.filter(r => 
-          r.name.toLowerCase().includes(validatedArgs.repository_name!.toLowerCase()) ||
-          r.full_name.toLowerCase().includes(validatedArgs.repository_name!.toLowerCase())
-        );
-        mockData.repositories = filteredRepos;
-        mockData.total_repositories = filteredRepos.length;
-      }
-
-      // Remove stats if not requested
-      if (!validatedArgs.include_stats) {
-        mockData.repositories = mockData.repositories.map(r => ({
-          name: r.name,
-          full_name: r.full_name,
-          description: r.description,
-          language: r.language,
-          topics: r.topics,
-        })) as any;
-      }
-
-      return {
-        content: [
-          {
-            type: "text",
-            text: JSON.stringify(mockData, null, 2),
-          },
-        ],
-      };
-    }
-
-    // Process the actual data if available
-    let processedData = result;
-    
-    if (validatedArgs.repository_name) {
-      // Filter by repository name if specified
-      if (processedData.repositories && Array.isArray(processedData.repositories)) {
-        processedData.repositories = processedData.repositories.filter((r: any) =>
-          r.name?.toLowerCase().includes(validatedArgs.repository_name!.toLowerCase()) ||
-          r.full_name?.toLowerCase().includes(validatedArgs.repository_name!.toLowerCase())
-        );
-        processedData.total_repositories = processedData.repositories.length;
-      }
-    }
-
-    if (!validatedArgs.include_stats && processedData.repositories) {
-      // Remove stats fields if not requested
-      processedData.repositories = processedData.repositories.map((r: any) => ({
-        name: r.name,
-        full_name: r.full_name,
-        description: r.description,
-        language: r.language,
-        topics: r.topics,
-      }));
-    }
-
-    if (!validatedArgs.include_contributors && processedData.repositories) {
-      // Remove contributors if not requested
-      processedData.repositories = processedData.repositories.map((r: any) => {
-        const { contributors, ...rest } = r;
-        return rest;
-      });
-    }
-
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify({
-            ...processedData,
-            source_endpoint: endpoint_used,
-          }, null, 2),
-        },
-      ],
-    };
-  } catch (error: any) {
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify({
-            error: "Failed to fetch repository data",
-            message: error.message,
-            timestamp: new Date().toISOString(),
-          }, null, 2),
-        },
-      ],
-      isError: true,
-    };
-  }
-}
