--- mcp-elizaos-data/src/tools/get-summary-data.ts
@@ -1,290 +1,0 @@
-import { z } from "zod";
-import { fetchData } from "../http-client.js";
-
-// Input schema
-export const GetSummaryDataArgsSchema = z.object({
-  summary_type: z.enum(["overview", "statistics", "activity", "contributors"]).optional().default("overview").describe("Type of summary data to fetch"),
-  time_period: z.enum(["week", "month", "quarter", "year", "all"]).optional().default("month").describe("Time period for the summary"),
-  include_trends: z.boolean().optional().default(false).describe("Include trend analysis"),
-});
-
-// Output schema
-const ContributorSummarySchema = z.object({
-  total_contributors: z.number().describe("Total number of contributors"),
-  active_contributors: z.number().describe("Active contributors in time period"),
-  top_contributors: z.array(z.object({
-    login: z.string().describe("GitHub username"),
-    contributions: z.number().describe("Number of contributions"),
-    repositories: z.number().describe("Number of repositories contributed to"),
-  })).describe("Top contributors"),
-});
-
-const ActivitySummarySchema = z.object({
-  total_commits: z.number().describe("Total commits"),
-  total_issues: z.number().describe("Total issues"),
-  total_pull_requests: z.number().describe("Total pull requests"),
-  repositories_updated: z.number().describe("Number of repositories updated"),
-  plugin_releases: z.number().describe("Number of plugin releases"),
-});
-
-const StatisticsSummarySchema = z.object({
-  repositories: z.object({
-    total: z.number().describe("Total repositories"),
-    active: z.number().describe("Active repositories"),
-    archived: z.number().describe("Archived repositories"),
-  }).describe("Repository statistics"),
-  plugins: z.object({
-    total: z.number().describe("Total plugins"),
-    active: z.number().describe("Active plugins"),
-    experimental: z.number().describe("Experimental plugins"),
-    deprecated: z.number().describe("Deprecated plugins"),
-  }).describe("Plugin statistics"),
-  community: z.object({
-    total_stars: z.number().describe("Total stars across all repos"),
-    total_forks: z.number().describe("Total forks across all repos"),
-    total_contributors: z.number().describe("Total unique contributors"),
-  }).describe("Community statistics"),
-});
-
-const TrendDataSchema = z.object({
-  commits_trend: z.array(z.object({
-    date: z.string().describe("Date"),
-    count: z.number().describe("Number of commits"),
-  })).describe("Commit trend data"),
-  contributors_trend: z.array(z.object({
-    date: z.string().describe("Date"),
-    count: z.number().describe("Number of active contributors"),
-  })).describe("Contributors trend data"),
-  stars_trend: z.array(z.object({
-    date: z.string().describe("Date"),
-    count: z.number().describe("Total stars"),
-  })).describe("Stars trend data"),
-});
-
-const SummaryDataResponseSchema = z.object({
-  summary_type: z.string().describe("Type of summary"),
-  time_period: z.string().describe("Time period covered"),
-  generated_at: z.string().describe("When the summary was generated"),
-  overview: z.object({
-    ecosystem_health: z.enum(["excellent", "good", "moderate", "needs_attention"]).describe("Overall ecosystem health"),
-    recent_highlights: z.array(z.string()).describe("Recent notable achievements"),
-    key_metrics: z.object({
-      total_repositories: z.number(),
-      total_plugins: z.number(),
-      total_contributors: z.number(),
-      total_stars: z.number(),
-    }).describe("Key ecosystem metrics"),
-  }).optional().describe("Overview data"),
-  statistics: StatisticsSummarySchema.optional().describe("Detailed statistics"),
-  activity: ActivitySummarySchema.optional().describe("Activity summary"),
-  contributors: ContributorSummarySchema.optional().describe("Contributor summary"),
-  trends: TrendDataSchema.optional().describe("Trend analysis"),
-});
-
-export type GetSummaryDataArgs = z.infer<typeof GetSummaryDataArgsSchema>;
-export type SummaryDataResponse = z.infer<typeof SummaryDataResponseSchema>;
-
-// Tool schema for MCP
-export const getSummaryDataSchema = {
-  name: "get_summary_data",
-  description: "Fetch comprehensive summary data about the ElizaOS ecosystem including statistics, activity, and trends",
-  inputSchema: GetSummaryDataArgsSchema,
-};
-
-export async function getSummaryDataHandler(args: any) {
-  const validatedArgs = GetSummaryDataArgsSchema.parse(args);
-  
-  try {
-    // Try multiple potential endpoints for summary data
-    const endpoints = [
-      "/summaries/ecosystem-summary.json",
-      "/summaries/overview.json",
-      "/dump/summary.json",
-      "/elizaos_elizaos.github.io/summary.json",
-    ];
-
-    let result = null;
-    let endpoint_used = "";
-
-    for (const endpoint of endpoints) {
-      const response = await fetchData<any>(endpoint);
-      if (response.success && response.data) {
-        result = response.data;
-        endpoint_used = endpoint;
-        break;
-      }
-    }
-
-    if (!result) {
-      // Create mock summary data if no real endpoint is available
-      const now = new Date();
-      const generateTrendData = (baseCount: number, days: number = 30) => {
-        return Array.from({ length: days }, (_, i) => {
-          const date = new Date(now.getTime() - (days - i - 1) * 24 * 60 * 60 * 1000);
-          const variance = Math.random() * 0.3 - 0.15; // ±15% variance
-          return {
-            date: date.toISOString().split('T')[0],
-            count: Math.max(0, Math.floor(baseCount * (1 + variance)))
-          };
-        });
-      };
-
-      const mockData: any = {
-        summary_type: validatedArgs.summary_type,
-        time_period: validatedArgs.time_period,
-        generated_at: now.toISOString(),
-      };
-
-      // Add data based on requested summary type
-      switch (validatedArgs.summary_type) {
-        case "overview":
-          mockData.overview = {
-            ecosystem_health: "good",
-            recent_highlights: [
-              "Released MCP integration plugin",
-              "Added 3 new blockchain plugins",
-              "Reached 3000+ GitHub stars",
-              "Community contributions increased 40%"
-            ],
-            key_metrics: {
-              total_repositories: 4,
-              total_plugins: 8,
-              total_contributors: 47,
-              total_stars: 3256,
-            }
-          };
-          break;
-
-        case "statistics":
-          mockData.statistics = {
-            repositories: {
-              total: 4,
-              active: 4,
-              archived: 0,
-            },
-            plugins: {
-              total: 8,
-              active: 6,
-              experimental: 2,
-              deprecated: 0,
-            },
-            community: {
-              total_stars: 3256,
-              total_forks: 734,
-              total_contributors: 47,
-            }
-          };
-          break;
-
-        case "activity":
-          const periodMultiplier = {
-            week: 1,
-            month: 4,
-            quarter: 12,
-            year: 52,
-            all: 100
-          }[validatedArgs.time_period];
-
-          mockData.activity = {
-            total_commits: 245 * periodMultiplier,
-            total_issues: 23 * periodMultiplier,
-            total_pull_requests: 67 * periodMultiplier,
-            repositories_updated: 4,
-            plugin_releases: 3 * Math.ceil(periodMultiplier / 4),
-          };
-          break;
-
-        case "contributors":
-          mockData.contributors = {
-            total_contributors: 47,
-            active_contributors: validatedArgs.time_period === "week" ? 8 : 
-                               validatedArgs.time_period === "month" ? 15 : 
-                               validatedArgs.time_period === "quarter" ? 25 : 35,
-            top_contributors: [
-              { login: "shawmakesmagic", contributions: 245, repositories: 3 },
-              { login: "elizaos-core", contributions: 189, repositories: 4 },
-              { login: "plugin-maintainer", contributions: 156, repositories: 2 },
-              { login: "community-dev", contributions: 134, repositories: 3 },
-              { login: "ai-researcher", contributions: 98, repositories: 2 },
-            ]
-          };
-          break;
-      }
-
-      // Add trend data if requested
-      if (validatedArgs.include_trends) {
-        const trendDays = {
-          week: 7,
-          month: 30,
-          quarter: 90,
-          year: 365,
-          all: 365
-        }[validatedArgs.time_period];
-
-        mockData.trends = {
-          commits_trend: generateTrendData(8, trendDays),
-          contributors_trend: generateTrendData(3, trendDays),
-          stars_trend: generateTrendData(15, trendDays).map((item, index) => ({
-            ...item,
-            count: 2500 + index * 2 + Math.floor(Math.random() * 10)
-          }))
-        };
-      }
-
-      mockData.source = "Generated summary data";
-
-      return {
-        content: [
-          {
-            type: "text",
-            text: JSON.stringify(mockData, null, 2),
-          },
-        ],
-      };
-    }
-
-    // Process the actual data if available
-    let processedData = result;
-    
-    // Filter data based on summary type
-    if (validatedArgs.summary_type !== "overview") {
-      const allowedFields = ["summary_type", "time_period", "generated_at", validatedArgs.summary_type];
-      if (validatedArgs.include_trends) {
-        allowedFields.push("trends");
-      }
-      
-      processedData = Object.keys(processedData)
-        .filter(key => allowedFields.includes(key))
-        .reduce((obj: any, key) => {
-          obj[key] = processedData[key];
-          return obj;
-        }, {});
-    }
-
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify({
-            ...processedData,
-            source_endpoint: endpoint_used,
-          }, null, 2),
-        },
-      ],
-    };
-  } catch (error: any) {
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify({
-            error: "Failed to fetch summary data",
-            message: error.message,
-            timestamp: new Date().toISOString(),
-          }, null, 2),
-        },
-      ],
-      isError: true,
-    };
-  }
-}
