--- mcp-elizaos-data/src/tools/list-available-plugins.ts
@@ -1,299 +1,0 @@
-import { z } from "zod";
-import { fetchData } from "../http-client.js";
-
-// Input schema
-export const ListAvailablePluginsArgsSchema = z.object({
-  category: z.string().optional().describe("Filter plugins by category (e.g., 'blockchain', 'social', 'ai')"),
-  include_details: z.boolean().optional().default(false).describe("Include detailed plugin information"),
-  sort_by: z.enum(["name", "category", "updated", "popularity"]).optional().default("name").describe("Sort plugins by field"),
-});
-
-// Output schema
-const PluginInfoSchema = z.object({
-  name: z.string().describe("Plugin name"),
-  category: z.string().optional().describe("Plugin category"),
-  description: z.string().optional().describe("Plugin description"),
-  version: z.string().optional().describe("Plugin version"),
-  author: z.string().optional().describe("Plugin author"),
-  repository: z.string().optional().describe("Plugin repository URL"),
-  documentation: z.string().optional().describe("Documentation URL"),
-  dependencies: z.array(z.string()).optional().describe("Plugin dependencies"),
-  keywords: z.array(z.string()).optional().describe("Plugin keywords"),
-  last_updated: z.string().optional().describe("Last update timestamp"),
-  status: z.enum(["active", "deprecated", "experimental"]).optional().describe("Plugin status"),
-});
-
-const PluginListResponseSchema = z.object({
-  total_plugins: z.number().describe("Total number of plugins"),
-  plugins: z.array(PluginInfoSchema).describe("Array of plugin information"),
-  categories: z.array(z.string()).describe("Available plugin categories"),
-  updated_at: z.string().describe("When the data was last updated"),
-});
-
-export type ListAvailablePluginsArgs = z.infer<typeof ListAvailablePluginsArgsSchema>;
-export type PluginListResponse = z.infer<typeof PluginListResponseSchema>;
-
-// Tool schema for MCP
-export const listAvailablePluginsSchema = {
-  name: "list_available_plugins",
-  description: "List all available plugins from ElizaOS with optional filtering and sorting capabilities",
-  inputSchema: ListAvailablePluginsArgsSchema,
-};
-
-export async function listAvailablePluginsHandler(args: any) {
-  const validatedArgs = ListAvailablePluginsArgsSchema.parse(args);
-  
-  try {
-    // Try multiple potential endpoints for plugin listings
-    const endpoints = [
-      "/elizaos-plugins_registry/plugins.json",
-      "/elizaos-plugins_registry/index.json",
-      "/summaries/plugin-list.json",
-      "/dump/available-plugins.json",
-    ];
-
-    let result = null;
-    let endpoint_used = "";
-
-    for (const endpoint of endpoints) {
-      const response = await fetchData<any>(endpoint);
-      if (response.success && response.data) {
-        result = response.data;
-        endpoint_used = endpoint;
-        break;
-      }
-    }
-
-    if (!result) {
-      // Create mock data structure if no real endpoint is available
-      const mockPlugins = [
-        {
-          name: "plugin-evm",
-          category: "blockchain",
-          description: "Ethereum Virtual Machine integration for ElizaOS agents",
-          version: "1.2.0",
-          author: "ElizaOS Team",
-          repository: "https://github.com/elizaos/elizaos-plugins/tree/main/plugin-evm",
-          documentation: "https://elizaos.github.io/docs/plugins/evm",
-          dependencies: ["ethers", "web3"],
-          keywords: ["ethereum", "blockchain", "web3", "smart-contracts"],
-          last_updated: "2024-01-15T10:30:00Z",
-          status: "active" as const
-        },
-        {
-          name: "plugin-solana",
-          category: "blockchain", 
-          description: "Solana blockchain integration for trading and NFT operations",
-          version: "1.1.5",
-          author: "ElizaOS Team",
-          repository: "https://github.com/elizaos/elizaos-plugins/tree/main/plugin-solana",
-          documentation: "https://elizaos.github.io/docs/plugins/solana",
-          dependencies: ["@solana/web3.js", "@solana/spl-token"],
-          keywords: ["solana", "blockchain", "nft", "trading"],
-          last_updated: "2024-01-14T15:45:00Z",
-          status: "active" as const
-        },
-        {
-          name: "plugin-twitter",
-          category: "social",
-          description: "Twitter/X integration for social media automation",
-          version: "2.0.1",
-          author: "ElizaOS Team",
-          repository: "https://github.com/elizaos/elizaos-plugins/tree/main/plugin-twitter",
-          documentation: "https://elizaos.github.io/docs/plugins/twitter",
-          dependencies: ["twitter-api-v2"],
-          keywords: ["twitter", "social", "automation", "posting"],
-          last_updated: "2024-01-16T09:20:00Z",
-          status: "active" as const
-        },
-        {
-          name: "plugin-farcaster",
-          category: "social",
-          description: "Farcaster protocol integration for decentralized social",
-          version: "0.9.2",
-          author: "Community",
-          repository: "https://github.com/elizaos/elizaos-plugins/tree/main/plugin-farcaster",
-          documentation: "https://elizaos.github.io/docs/plugins/farcaster",
-          dependencies: ["@farcaster/js"],
-          keywords: ["farcaster", "decentralized", "social", "protocol"],
-          last_updated: "2024-01-12T14:10:00Z",
-          status: "experimental" as const
-        },
-        {
-          name: "plugin-coingecko",
-          category: "data",
-          description: "CoinGecko API integration for cryptocurrency price data",
-          version: "1.0.3",
-          author: "ElizaOS Team",
-          repository: "https://github.com/elizaos/elizaos-plugins/tree/main/plugin-coingecko",
-          documentation: "https://elizaos.github.io/docs/plugins/coingecko",
-          dependencies: ["axios"],
-          keywords: ["cryptocurrency", "prices", "data", "coingecko"],
-          last_updated: "2024-01-13T11:25:00Z",
-          status: "active" as const
-        },
-        {
-          name: "plugin-knowledge",
-          category: "ai",
-          description: "Knowledge base and RAG capabilities for enhanced AI responses",
-          version: "1.3.0",
-          author: "ElizaOS Team",
-          repository: "https://github.com/elizaos/elizaos-plugins/tree/main/plugin-knowledge",
-          documentation: "https://elizaos.github.io/docs/plugins/knowledge",
-          dependencies: ["vector-db", "embeddings"],
-          keywords: ["rag", "knowledge", "ai", "embeddings"],
-          last_updated: "2024-01-15T16:40:00Z",
-          status: "active" as const
-        },
-        {
-          name: "plugin-mcp",
-          category: "integration",
-          description: "Model Context Protocol integration for external tool access",
-          version: "1.0.0",
-          author: "ElizaOS Team",
-          repository: "https://github.com/elizaos/elizaos-plugins/tree/main/plugin-mcp",
-          documentation: "https://elizaos.github.io/docs/plugins/mcp",
-          dependencies: ["@modelcontextprotocol/sdk"],
-          keywords: ["mcp", "tools", "integration", "protocol"],
-          last_updated: "2024-01-16T08:15:00Z",
-          status: "active" as const
-        },
-        {
-          name: "plugin-auton8n",
-          category: "automation",
-          description: "Advanced automation workflows and task orchestration",
-          version: "0.8.1",
-          author: "Community",
-          repository: "https://github.com/elizaos/elizaos-plugins/tree/main/plugin-auton8n",
-          documentation: "https://elizaos.github.io/docs/plugins/auton8n",
-          dependencies: ["workflow-engine"],
-          keywords: ["automation", "workflows", "orchestration", "tasks"],
-          last_updated: "2024-01-11T13:30:00Z",
-          status: "experimental" as const
-        }
-      ];
-
-      // Filter by category if specified
-      let filteredPlugins = mockPlugins;
-      if (validatedArgs.category) {
-        filteredPlugins = mockPlugins.filter(p => 
-          p.category?.toLowerCase() === validatedArgs.category!.toLowerCase()
-        );
-      }
-
-      // Sort plugins
-      const sortField = validatedArgs.sort_by;
-      filteredPlugins.sort((a, b) => {
-        switch (sortField) {
-          case "name":
-            return a.name.localeCompare(b.name);
-          case "category":
-            return (a.category || "").localeCompare(b.category || "");
-          case "updated":
-            return new Date(b.last_updated || 0).getTime() - new Date(a.last_updated || 0).getTime();
-          case "popularity":
-            // Mock popularity based on name (just for demo)
-            const popularityOrder = ["plugin-twitter", "plugin-evm", "plugin-solana", "plugin-knowledge"];
-            return popularityOrder.indexOf(a.name) - popularityOrder.indexOf(b.name);
-          default:
-            return 0;
-        }
-      });
-
-      // Remove detailed fields if not requested
-      if (!validatedArgs.include_details) {
-        filteredPlugins = filteredPlugins.map(p => ({
-          name: p.name,
-          category: p.category,
-          description: p.description,
-          version: p.version,
-          status: p.status,
-        })) as any;
-      }
-
-      const categories = [...new Set(mockPlugins.map(p => p.category).filter(Boolean))];
-
-      const mockData = {
-        total_plugins: filteredPlugins.length,
-        plugins: filteredPlugins,
-        categories,
-        updated_at: new Date().toISOString(),
-        source: "Generated from available plugin directories"
-      };
-
-      return {
-        content: [
-          {
-            type: "text",
-            text: JSON.stringify(mockData, null, 2),
-          },
-        ],
-      };
-    }
-
-    // Process the actual data if available
-    let processedData = result;
-    
-    if (validatedArgs.category && processedData.plugins) {
-      // Filter by category if specified
-      processedData.plugins = processedData.plugins.filter((p: any) =>
-        p.category?.toLowerCase() === validatedArgs.category!.toLowerCase()
-      );
-      processedData.total_plugins = processedData.plugins.length;
-    }
-
-    if (!validatedArgs.include_details && processedData.plugins) {
-      // Remove detailed fields if not requested
-      processedData.plugins = processedData.plugins.map((p: any) => ({
-        name: p.name,
-        category: p.category,
-        description: p.description,
-        version: p.version,
-        status: p.status,
-      }));
-    }
-
-    // Sort plugins if data exists
-    if (processedData.plugins && validatedArgs.sort_by) {
-      const sortField = validatedArgs.sort_by;
-      processedData.plugins.sort((a: any, b: any) => {
-        switch (sortField) {
-          case "name":
-            return (a.name || "").localeCompare(b.name || "");
-          case "category":
-            return (a.category || "").localeCompare(b.category || "");
-          case "updated":
-            return new Date(b.last_updated || 0).getTime() - new Date(a.last_updated || 0).getTime();
-          default:
-            return 0;
-        }
-      });
-    }
-
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify({
-            ...processedData,
-            source_endpoint: endpoint_used,
-          }, null, 2),
-        },
-      ],
-    };
-  } catch (error: any) {
-    return {
-      content: [
-        {
-          type: "text",
-          text: JSON.stringify({
-            error: "Failed to fetch plugin list",
-            message: error.message,
-            timestamp: new Date().toISOString(),
-          }, null, 2),
-        },
-      ],
-      isError: true,
-    };
-  }
-}
